validacia pre modelov

+++ app/controllers/districts.controller.js
@@ -13,6 +13,17 @@ exports.create = (req, res) => {
         return;
     }
 
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const districts = {
         name: req.body.name,
         vehRegNum: req.body.vehRegNum,
@@ -36,6 +47,17 @@ exports.create = (req, res) => {
 };
 
 exports.delete = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Districts.destroy({
@@ -64,6 +86,17 @@ exports.delete = (req, res) => {
 };
 
 exports.update = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Districts.update(req.body, {
@@ -92,6 +125,17 @@ exports.update = (req, res) => {
 };
 
 exports.get = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Districts.findByPk(id)
@@ -118,6 +162,17 @@ exports.get = (req, res) => {
 };
 
 exports.getAll = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const pageNumber = req.params.pageNumber;
     const pageSize = req.params.pageSize;
