+++ app/controllers/villages.controller.js
@@ -13,6 +13,17 @@ exports.create = (req, res) => {
         return;
     }
 
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const villages = {
         fullName: req.body.fullName,
         shortName: req.body.shortName,
@@ -37,6 +48,17 @@ exports.create = (req, res) => {
 };
 
 exports.delete = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Villages.destroy({
@@ -65,6 +87,17 @@ exports.delete = (req, res) => {
 };
 
 exports.update = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Villages.update(req.body, {
@@ -93,6 +126,17 @@ exports.update = (req, res) => {
 };
 
 exports.get = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const id = req.params.id;
 
     Villages.findByPk(id)
@@ -119,6 +163,17 @@ exports.get = (req, res) => {
 };
 
 exports.getAll = (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).send({
+            timestamp: new Date().toISOString(),
+            message: "Údaje boli nesprávne vyplnené",
+            error: true,
+            validations:  errors.array(),
+            nav: req.protocol + '://' + req.get('host')
+        });
+    }
+
     const pageNumber = req.params.pageNumber;
     const pageSize = req.params.pageSize;
